+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
 
-* 2019/5/14: Initial public release - [@dblock](https://github.com/dblock).
+* 2020/04/23: Added `set sender` to change which Slack admin sends invitations - [@dblock](https://github.com/dblock).
+* 2019/05/14: Initial public release - [@dblock](https://github.com/dblock).